tests: merge end2end tests with anacondalib testcase class
Move missing functions from end2end to anacondalib
Rework StorageEncryption e2e test to TestStorageEncryption
Add new test_cases directory for storing tplib files
Remove end2end tests
