+++ test/anacondalib.py
@@ -17,6 +17,7 @@
 
 import os
 import sys
+from subprocess import CalledProcessError
 
 # import Cockpit's machinery for test VMs and its browser test API
 TEST_DIR = os.path.dirname(__file__)
@@ -26,9 +27,12 @@ sys.path.append(os.path.join(os.path.dirname(TEST_DIR), "bots/machine"))
 
 from machine_install import VirtInstallMachine
 from testlib import MachineCase  # pylint: disable=import-error
+from utils import add_public_key
 
 from storage import Storage
 from language import Language
+from progress import Progress
+
 
 pixel_tests_ignore = [".logo", "#betanag-icon"]
 
@@ -40,6 +44,7 @@ class VirtInstallMachineCase(MachineCase):
     @classmethod
     def setUpClass(cls):
         VirtInstallMachine.efi = cls.efi
+        cls.ext_logging = bool(int(os.environ.get('EXTENDED_LOGGING', '0')))
 
     def setUp(self):
         # FIXME: running this in destructive tests fails because the SSH session closes before this is run
@@ -50,6 +55,7 @@ class VirtInstallMachineCase(MachineCase):
         super().setUp()
 
         self.allow_journal_messages('.*cockpit.bridge-WARNING: Could not start ssh-agent.*')
+        self.installation_finished = False
 
     def resetLanguage(self):
         m = self.machine
@@ -70,3 +76,40 @@ class VirtInstallMachineCase(MachineCase):
         # CLEAR_PARTITIONS_DEFAULT = -1
         s.dbus_set_initialization_mode(-1)
         s.dbus_scan_devices()
+
+    def downloadLogs(self):
+        if not self.ext_logging:
+            return
+
+        self.logs_dir = os.path.join('./test_logs', f'{self.__class__.__name__}.{self._testMethodName}')
+        if not os.path.isdir(self.logs_dir):
+            os.makedirs(self.logs_dir)
+
+        self.machine.download('/tmp/anaconda.log', 'anaconda.log', self.logs_dir)
+        self.machine.download('/tmp/packaging.log', 'packaging.log', self.logs_dir)
+        self.machine.download('/tmp/storage.log', 'storage.log', self.logs_dir)
+        self.machine.download('/tmp/dbus.log', 'dbus.log', self.logs_dir)
+        self.machine.download('/tmp/syslog', 'syslog', self.logs_dir)
+        try:
+            self.machine.download('/tmp/anaconda-tb-*', '.', self.logs_dir)
+        except CalledProcessError:
+            pass
+
+    def handleReboot(self):
+        """
+        Method for rebooting into the installed system.
+
+        Should be called when installation is finished. Performs necessary steps
+        so the test can continue with checking installed system.
+        """
+        add_public_key(self.machine)
+        self.downloadLogs()
+        self.installation_finished = True
+        p = Progress(self.browser)
+        p.reboot()
+        self.machine.wait_reboot()
+
+    def tearDown(self):
+        if not self.installation_finished:
+            self.downloadLogs()
+        super().tearDown()
