--- test/end2end/github_pr.plan.yaml
@@ -1,13 +1,0 @@
-name: WebUI e2e on PR
-description: Runs all WebUI end to end tests on pull request
-point_person: rvykydal@redhat.com
-artifact_type: github.pr.anaconda
-verified_by:
-  test_cases:
-    query: '"anaconda-webui" == tc.execution.type and "disabled" not in tc.tags'
-configurations:
-  - architecture: x86_64
-    branch: master
-reporting:
-  - type: github-pr
-  - type: xunit
