--- test/end2end/wizard_navigation.tc.yaml
@@ -1,27 +1,0 @@
-name: Wizard navigation
-description: |
-  Tries various navigation methods in the installation wizard and makes sure
-  the installation can be finished
-author: zveleba@redhat.com
-tags:
-  - anaconda
-priority: 5
-execution:
-  type: anaconda-webui
-  automation_data:
-    script_file: ./ui/webui/test/end2end/wizard_navigation.py
-    test_case: WizardNavigation
-instructions:
-  setup:
-    - Start installation in VM
-  steps:
-    - step: Go through all the required steps and stop at the review screen.
-    - step: Use the back button until you get to the disk selection screen.
-    - step: |
-        Do the neccesary steps to get to the review screen again and start
-        the installation
-      result: Installation finishes without problems
-    - step: Reboot
-      result: Installed system starts
-  teardown:
-    - step: Remove VM
