+++ test/test_cases/default.tc.yaml
@@ -8,8 +8,8 @@ priority: 9
 execution:
   type: anaconda-webui
   automation_data:
-    script_file: ./ui/webui/test/end2end/default.py
-    test_case: DefaultInstallation
+    script_file: ./test/check-progress
+    test_case: TestInstallationProgress.testBasic
 instructions:
   setup:
     - Start installation in VM
