+++ test/test_cases/periodic.plan.yaml
@@ -1,5 +1,5 @@
-name: WebUI end2end periodic tests
-description: Runs all WebUI end2end tests periodically
+name: WebUI periodic tests
+description: Runs all WebUI tests periodically
 point_person: rvykydal@redhat.com
 artifact_type: github.webui.e2e.scheduled
 verified_by:
